requirements.txt
fastapi[all]
uvicorn[standard]
pydantic-settings
email-validator
python-multipart
python-dotenv
pdfplumber
python-docx
langchain
langchain-openai
openai
tiktoken
python-magic
supabase
logtail-python
passlib[bcrypt]
zxcvbn
phonenumbers